autohdr_exe/requirements.txt
customtkinter==5.2.2
requests==2.31.0
tkinterdnd2-universal==1.7.3
python-dotenv==1.0.1
